Add support for Lifecast LDI3 format
