+++ web/embed_in_div.html
@@ -8,9 +8,9 @@
 <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
 
 <script type="module">
-  import * as LifecastVideoPlayer from './lifecast_res/LifecastVideoPlayer8.js';
+  import * as LifecastVideoPlayer from './lifecast_res/LifecastVideoPlayer9.js';
   LifecastVideoPlayer.init({
-    _media_url: "media/000000.jpg",
+    _media_url: "media/2layer.jpg",
     _embed_in_div: "player_div"
   });
 </script>
