+++ web/index.html
@@ -8,8 +8,8 @@
 <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
 
 <script type="module">
-  import * as LifecastVideoPlayer from './lifecast_res/LifecastVideoPlayer8.js';
-  LifecastVideoPlayer.init({_media_url: "media/000000.jpg"});
+  import * as LifecastVideoPlayer from './lifecast_res/LifecastVideoPlayer9.js';
+  LifecastVideoPlayer.init({_media_url: "media/2layer.jpg"});
 </script>
 
 </head>
